extended margins of the calculation

+++ src/main/java/com/ivan/service/FibonacciServiceImpl.java
@@ -50,13 +50,13 @@ public class FibonacciServiceImpl implements FibonacciService {
         return fibRepository.findAllIndexes();
     }
 
-    private int fibonacci(int n) {
+    private long fibonacci(int n) {
         if (n == 0) return 0;
         if (n == 1) return 1;
 
-        int prevPrev = 0;
-        int prev = 1;
-        int result = 0;
+        long prevPrev = 0;
+        long prev = 1;
+        long result = 0;
 
         for (int i = 2; i <= n; i++)
         {
